pyproject.toml
[project]
name = "yellow-pages"
version = "0.1.0"
description = "MCP server that turns natural language into API calls via RAG over OpenAPI schemas"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "requests>=2.31.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]

